Remove multiline support and SASS checker [#11]
Multiline support broke error messages for various other checkers.
Remove this support for now, since only the SASS checker needs it.
Remove this checker, too, because it is broken by this change.

Once we got rid of flymake, we'll have proper support for multiline
error messages, and can re-introduce this checker.