Fix duplicate error messages [#96]
Re-use existing overlays for errors to avoid duplicate overlays and
error messages.  This would happen with chained checkers.  After the
last checker was run, *all* errors are reported again, including those
of the previous checkers.  Hence overlays for errors from previous
checkers would be added again and again.  While this doesn't matter with
regards to highlighting, it resulted in duplicate error messages when
showing all error messages as implemented in #96, which is really
annoying.