Refactor test suite to ease test running
Merge testhelpers back into testsuite files to reduce load hassle, and
put all testsuite symbols into the `flycheck-testsuite` namespace.

The testsuite.el does not longer change global settings, e.g. like
`debug-on-error`, and emits no messages anymore, so it's clean to load
into a running Emacs.

Add a test runner script which takes a test selector as argument so that
we don't need to run the whole test suite each time, which is a huge
relief now that we have well over hundred test cases.  The whole thing
is wrapped up as `tests/run.sh` which is now the preferred way to run
tests interactively on the terminal.