Refactor Google This usage
As `google-string` is autoloadable in recent Google This versions,
just check it is `fbound`  instead of trying to require the feature.

Wrap the use of the function into an if expression, so that the byte
compiler recognizes the conditional use of the function and does not
emit a warning anymore.