Pseudo-bootstrappify the manual :)
I'd like to use Bootstrap directly, but getting Texinfo to output HTML 5
with the right elements for Bootstrap is a major PITA, and requires
advanced Perl wizardry.  That's definitely beyond my limits, so I just
copied the corresponding styles from the Bootstrap stylesheets.  The
results are not ideal, but quite ok.