Bootstrap Puppet and all modules with Puppet
Use a Puppet manifest to Bootstrap Puppet 3 and the modules we need.
This gets rid of the additional shell script, and the requirement for
librarian-puppet.

Modules are now installed within the testing VM, without the need for
additional software, which makes VM creation a lot easier, and prevents
trouble as in #236.