Add options to flag C++ exception handling as an error.
This serves pretty much the same purpose as the options to flag C++ RTTI
as an error, being useful mostly in freestanding and/or embedded
environments where linking in the required runtime support is
infeasible. Adding an option to disable exception handling makes it
easier to catch a mistake early.