Extract error IDs in error patterns of checkers
Affects: pylint, flake8, chktex, rubocop, phpcs, perlcritic

Note that phpcs uses checkstyle parsing again.  This might revive the
output issues in #78 and #118, but we choose to ignore these from now
on.  It's not our fault that PHP is broken, and we'll not work around it
anymore.

Also, lower gsjlint level to warning, since it doesn't really report
errors.