[Fix #559] Don't cache last checker without using it
Only cache the last used syntax checker when actually using it, to make
flycheck-get-checker-for-buffer free of side effects.

Specifically Flycheck calls this function to determine whether it may
enable Flycheck Mode in a particular buffer when Global Flycheck Mode is
on.  In Emacs 24.3 and earlier, this happens at a point where hooks have
not yet run, which in turn affects the predicates of registered syntax
checkers, if they check for specific major modes.

As a result, a syntax checker would have been cached as last checker and
used for the buffer subsequently, even if a preferred checker was
available later, if a minor mode was enabled through hooks.