Made scss-lint install task install scss_lint
The scss-lint ruby gem has changed name to scss_lint (note the
underscore instead of a hyphen). The binary command (with the hyphen)
has stayed the same, but the name of the gem to install has changed.

See https://github.com/brigade/scss-lint/blob/master/CHANGELOG.md for
details

This commit just updates the install task to install the gem using the
new name.