Add support for JavaScript Standard Style
This commit introduces support for "standard", a tool to ensure code conforms
to the "JavaScript Standard Style" (https://github.com/feross/standard).

The checker is built upon eslint and is meant to be used as-is, without
configuration, hence it is not being chained to other checkers (unlike
jshint, eslint, etc).