Add c/c++-clang-check checker
The clang project provides a clang-check binary which can utilise a
compilation database. In a transitional Makefile based system you create
it with a tool called bear:

  bear make

Which will leave a compile_commands.json file in the root of your build.
When clang-check is called it will re-use the compile directives from
the json database. This eliminates the need to manually add include
patch to your flycheck setup.

As clang-check uses the path of the file to find the json database this
checker uses `source_original` and adds the :predicate
flycheck-buffer-saved-p.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>