Add a 'filter' command to the error list
Pressing "f" in the error list now prompts for an error level. All
errors whose severity level is below that level are hidden. Pressing "F"
resets the filter. The current filter is shown in the mode line. The
docs mention the two new commands.

Changing filter levels scrolls back to the top of the error list. This
is to avoid confusing behavior when filters are removed. Filtering out
the first entry of the error list, for example, moves the second entry
to the top. If we didn't scroll back up, removing the filter would leave
the first entry invisible, causing confusion.

Closes #698.