Add customizable function that wraps each checker command
The function is usefull to wrap a checker command, if the checker
is only available in sandbox-like enviroments. In NixOS, checkers
are usually only available in a project environement that can be
accessed with the "nix-shell" command. The function in this case
would simply wrap each checker command with "nix-shell --command ..."