Add stylelint checker for CSS, SCSS and Less
Define a checker for stylelint that parses JSON

The stylelint CLI can format its output as text or JSON. The text
formatter is not printing a severity keyword. That's why I used the JSON
formatter and added a parser for the JSON output.

See: https://stylelint.io/

Show CssSyntaxErrors with checker css-stylelint

When the CSS contains syntax errors and it is passed to the
css-stylelint checker, the error output does not contain JSON data and
therefore leads to a "checker probably flawed" error message. To avoid
this error message and to show the user where the CSS syntax error is,
the stylelint CssSyntaxError message is parsed and shown in the right
position.

I used the build in flycheck-parse-with-patterns and an error pattern in
:error-patterns to introduce as little new functionality as possible.

Show stylelint configuration errors in line 1

A problem with the stylelint configuration is not returned in JSON
format and has to be parsed with an error pattern aswell. The error
pattern does not contain a line number. Because errors without line
numbers can't be shown, an error-filter is needed to add a line number
of 1. This filter also prepends the error message with "Configuration
error: " to tell the user that the error occured because of a
misconfiguration of stylelint.

The flycheck-sanitize-errors filter is needed because the error-pattern
for CssSyntaxErrors returns messages with unnecessary whitespace.

Add config-file-var for stylelint checker

The stylelint library has a good lookup mechanism for the default
configuration, thats why the standard value of the config-file-var
flycheck-stylelintrc is set to nil.
See: http://stylelint.io/user-guide/configuration/#loading-configuration

Enable css-stylelint in scss-mode

stylelint can parse scss syntax (with postcss-scss) and can be used in
scss-mode aswell. To avoid having to specify an extra checker (and
duplicate some code), I decided to just activate the "--syntax" option
of stylelint. It should work with normal css files aswell.

Replace stylelint JSON parser with :error-patterns

The CLI output of stylelint was improved and shows error severity now.
To avoid adding custom code to flycheck the JSON parser can be replaced
with error-patterns. This is possible because of these two PRs:

- stylelint/stylelint#1023
  Show CssSyntaxErrors along the normal linter warnings (threw an
  Exception before)

- stylelint/stylelint#1056
  Show an icon for severity and show rule name in extra column

The :error-patterns cover the following scenarios:

- Linting errors of any severity
  The severity is shown with an icon from the npm package log-symbols.
  These vary depending on the operating system (different icons for
  windows). Thats why the error and warning pattern contain two
  different symbols.
  See: https://github.com/sindresorhus/log-symbols

  It is possible to use custom severities in stylelint. The info pattern
  matches all linting errors with a severity other than error or
  warning. Info seems to fit best because the user could use error or
  warning otherwise.

- Deprecation Warnings and Invalid Options
  The output also informs the user about deprecated rules and invalid
  configuration options. This helpful information is shown in the first
  line.

- Errors
  Some things result in a thrown Error. For example, a missing
  configuration file, errors in the YAML or JSON file, no "rules"
  property in the configuration, using non-existent rules. Especially
  missing configurations files or errors in the configuration file seem
  likely. To avoid a flycheck error ("checker definition probably
  flawed") in these cases, their error message is shown in the first
  line.

Add --quiet flag to css-stylelint syntax checker

Most warnings can be automatically fixed and some user might not want
to show them (the quiet flag only shows errors with severity "error").

Replace :error-patterns with :error-parser

All flycheck checkers should prefer structured output. That's why the
stylelint JSON formatter with a custom :error-parser is the recommended
way.

The JSON parser gets deprecation warnings, invalid option errors and
linting errors from stylelint and turns them into flycheck errors.

stylelint sometimes throws errors and does not output JSON. The error
parser handles this case and uses a regular expression to get an error
message.

Move stylelint for scss-mode in extra checker
Add a stylelint checker for less-css-mode

Save common args in a constant and use the same config and options
variable in all checkers to reduce duplication.