Set C locale & patch C, Haskell and fortran tests
gcc, ghc and gfortran output smartquotes in a Unicode locale and pure
ASCII in the C locale.  Flycheck's test suite assumes a Unicode locale,
so the test suite can unexpectedly fail when only the C locale is
available.

This patch sets the test suite to always use the C locale and updates
expected gcc, ghc and gfortran output accordingly.  This makes the test
suite more portable, in particular to the Debian package autobuilder
infrastructure.