set epg-gpg-home-directory
gpg immediately dies with an error when invoked if the current user's
home directory does not exist (as is the case on a Debian package
autobuilder).  This causes all tests that invoke gpg to be skipped.
Setting `epg-gpg-home-directory` when running the test suite fixes the
problem, making the test suite more portable.