Pass relevant settings to checkdoc subprocess
‘flycheck-emacs-lisp-checkdoc-variables’ contains all variables that are
passed to the inferior Emacs process running checkdoc if they are
already bound in the parent process. This list is kept explicitly
because generating it requires browsing the checkdoc customization group
which is only populated after loading checkdoc. A new test ensures that
the list is up-to-date.

The main use of this is allowing for file- and directory-local checkdoc
settings.

Fixes GH-741 and closes GH-937.