Add option to set Rust binary name
When there are multiple targets, we should call `cargo rustc --bin NAME`
with NAME being the binary name to build.

This new variable allows one to set the binary name.  rust-cargo will
only use it when `flycheck-rust-crate-type` is `bin`.  If it is unset,
we fallback to the previous behavior of not using `--bin`, to let cargo
build the default binary.

Closes GH-958