c/c++-cppcheck: Add support for multiple --std= arguments
Previously it  was only possible to  pass a single --std=  argument to
cppcheck.  It  turns  out   that  cppcheck  supports  multiple  --std=
arguments (ie,  --std=posix --std=c99),  which cannot  be passed  in a
comma separated  list. This  change adds  support for  multiple --std=
arguments  by making  the emacs  variable either  a nil  or a  list of
strings.

This      change     also      renames      the     variable      from
`-cppcheck-language-standard` to `-cppcheck-standards` since these are
not strictly language standards anymore.