Strawman: Dynamic adjustment based on buffers
This is a strawman (not a proper cleaned up commit), to enable
discussion on the best UX for supporting additional buffers. The code is
intentionally poorly formatted with dangling `(` and `)` so that the
diff is a bit smaller. In the proper PR we'll format properly.

This PR will, when given an error list that contains an error whose
source buffer does not match the currently-being-flychecked buffer:

* Add another column, "Buffer", to the tabulate columns (flycheck-error-list-format).
* Re-generate that tabulation (`flycheck-retabulate`).
* Add that data to the output columns (`flycheck-error-list-make-entry).