Enable Racket checker when Geiser uses Racket
Geiser enhances scheme-mode's support for Racket as a minor mode
(geiser-mode). However, we cannot just check minor-mode-alist for
geiser-mode because Geiser supports other Scheme implementations as well
as Racket.

Jose Antonio Ortega Ruiz, the author of Geiser, has confirmed that
checking the geiser-impl--implementation variable is a public interface
for finding out what Scheme implementation Geiser is associating the
current buffer with: "--" in its name is just for suggesting that the
variable is not for an end user's interactive use.

The :modes declaration has some overlap in its functionality with the
:predicate declaration and yet is included because it will become
mandatory after all in the near future.

Closes GH-979