Add the ability to explain errors at point
This commit adds the `:error-explainer` checker property, an optional
function that can be used to produce an explanation message for a given
`flycheck-error` object.

The user can call a new interactive function
`flycheck-explain-error-at-point` to display an explanation for the
error under the cursor using the `:error-explainer` function of the
error's checker.  The error list also gained support for the
explanation: clicking on the error ID displays the explanation.

This commits adds initial error explainer functions to the `rust` and
`rust-cargo` checkers using `rustc --explain`.