Fix macro errors in JSON output for rust checkers
Macro expansion errors in the JSON output for the rust checkers use a
slightly different format than other errors.  As a result, the current
parser does not report macro expansion errors in flycheck, even though
they are present in the JSON output.

This commit enhances the parser for the rust JSON output by adding the
extra logic needed to capture the macro expansion errors.

Macro expansion errors have an 'expansion' field that contains another
'span', and this 'span' can contain an 'expansion', and so on.  So
in order to collect all the spans in a diagnostic, we need not only to
do a breadth search on the list of spans, but also a depth search on the
'expansion' field.

This search is delegated to a new function
`flycheck-parse-rust-collect-spans` that recursively collects all spans
in a 'root' span object.

Spans of macro expansion errors do not have any label, so we reuse the
text from the error message itself.

This commits also fixes one obsolete test for multi-line errors, as the
new error format actually makes this a single-line type error.  And it
adds a test for macro errors to ensure we don't regress on such cases in
the future.