Check whether buffers are really saved to disk
flycheck-buffer-saved-p checked if the current buffer visited a file,
but not if the file existed. So visiting a file that did not exist
returned true.
This caused breakages with checkers such as puppet-lint that don't take
input from stdin, used this predicate and expected a concrete file to
exist.
This commit makes flycheck-buffer-saved-p use file-exists-p to ensure
that the buffer visits a concrete file. It also adds tests for
flycheck-buffer-saved-p.

Fixes GH-1148