Handle errors from rustc that have no associated spans
Errors reported by rustc may not have any associated position in a file, for
instance if the error is related to the project as a whole (see
flycheck/flycheck-rust#47).

Previously, we did not collect any error without spans, thus not only
misreporting errors from rustc, but also raising a suspicious checker state.

This commit enhances `flycheck-parse-rust` to collect errors without associated
spans.  By default, these errors are bound to the position (0,0) of the current
buffer.