Fix filtering for overlay
Previously, buffers were checked for equality; now path names are
checked.

The issue is that flycheck-error-buffer will always return the buffer
that the flycheck checker was started with.  Now that
(flycheck-error-filename ...) is used in conjunction with (buffer-filename
(current-buffer)), the filtering is correct.

After this change, I'm only seeing the error overlays on the correct
locations.