Update flycheck-buttercup matchers to latest buttercup API
Buttercup has updated the `expect` macro to wrap its arguments in closures [1],
so we now have to `funcall` the arguments passed to `buttercup-define-matcher`
in order to obtain their actual value.

In addition, default matchers like `:to-be-truthy` or `:to-equal` modify
`match-data`, so we have to save the results of a regexp search ourselves.

[1]: https://github.com/jorgenschaefer/emacs-buttercup/commit/929a904debd4cf5846bbcb7e3c9ee21f7636c5bb