Fix search page not showing any results
I was a bit hasty in GH-1314.  The patch to prevent the analytics script to load
also disabled the search page and the readthedocs widget, since jQuery makes
heavy use of `document.createElement` as well.

This commit fixes that by overriding the HTMLScriptElement: when trying to load
something from the analytics server, the `src` attribute is silently discarded.
Otherwise, it runs as usual.