Fix rust-cargo for nightly rust
The cargo checker was not displaying any errors anymore, as it filtered them out
due to a path mismatch.

It looks like stable cargo returns filepaths relative to the CWD, whereas
nightly cargo returns filepaths relative to the manifest location.  So we can fix
the issue by using the manifest location as working directory for the checker,
and this should be backwards-compatible.

Closes GH-1378.