Test Emacs 26.0.91 on Travis
This required updating a few tests for elisp checkers, and one Cask
workaround.

The Cask workaround is to add `--verbose` to `cask install`.  See
cask/cask#367.  This is harmless for all Emacs versions.

On a related note, llvm-mode has been removed from MELPA because the
LLVM git repository is too large.  We'll have to fetch the mode directly
if we want to run all of our integration tests on CI.  I've opened
GH-1411 for that.

The JSON parser has become stricter in Emacs 26, and does not allow
newlines or tabs inside strings.  I fixed the test data that had those
invalid characters.

The check-declare output format has changed.  The column is now
optional.  This required yet another version guard in the test.  Not
sure this is not a regression though.

Paths returned by check-declare are now relative to the output; the root
file we used in the example file ended up with a bunch of "../../…".
Removing the `/` makes the test more robust.

For the emacs-lisp checker tests, `message` can now be called without
arguments.  so I've changed the tests to use another function.

Alos, the checker now emits a new error when compressed files do not
match the `provided` directive.  Using the correct name will make this
test retro-compatible.