Fix `rust-cargo` with Rust 1.26+
Due to a recent change to cargo [1], in a workspace setting (multiple
crates) filenames are now relative to the workspace root, whereas
previously they were relative to the crate.

This commit uses `cargo metadata` to figure out the workspace root, if
it exists.  If it doesn't, it fallbacks on the manifest directory.

[1]: https://github.com/rust-lang/cargo/pull/4788