Don't autoquote lambdas in `flycheck-define-checker`
In `flycheck-define-checker`, we function quoted (`#'`) all arguments
except `:command`.  This made for more readable checker definitions,
since we could pass lambdas or unquoted symbols to functions.

While this works in Flycheck itself, quoting lambdas fails when using
`with-eval-after-load`.  In the body of the latter, macros will be
expanded, and a `(lambda ...)` form is macro-expanded to
`(function (lambda ...)` resulting in a doubly-quoted lambda, which is
incorrect.

This commit solves the issue by not auto-quoting the arguments to
`flycheck-define-checker`.  For backward compatibility, it
function-quotes unquoted symbols.  This commit also updates all the
included checkers to use explicit function-quotes.

Fix GH-1398.