Add option `flycheck-jump-to-other-file-immediately`
This options controls whether Flycheck error navigation jumps to other
files immediately when the selected error is in another file. This is
related to PR #1427 that added overlays on the first line of the
current file when errors in other files prevent checking the current
file.

When the new option `flycheck-jump-to-other-file-immediately` is
nil (it defaults to non-nil), error navigation in the current file
moves the cursor to first-line overlay, and then you have to jump
again to actually go to the other file containing the error. The idea
is that it's surprising for Flycheck error navigation to change files,
and stopping on the first-line overlay first makes it less surprising.

There is more context for this commit here:
https://github.com/flycheck/flycheck/pull/1427#issuecomment-390912510.