Fix `flycheck-first-error` for first-line overlays
The first-line error overlays added in PR #1427 weren't handled
correctly by `flycheck-first-error`, due to a bug in
`flycheck-next-error-overlay-pos`, where that function assumed that
`point-min` would not be part of an error overlay in the `reset` case.

This bug and fix is unrelated to PR #1473, except that I ran into the
bug while testing #1473.