Fix `flycheck-next-error` for other-file overlays
When running `flycheck-next-error` on overlays for errors in other
files, we'd like to jump to the other file containing the
error. Before this commit, if the cursor was already on an other-file
overlay it would be skipped (at least for forward error searching).

I don't know a way to make the current file have both local errors and
overlays for errors in other files, so I did not test that case.