Allow checker executables to be set to a relative path
Previously, flycheck-CHECKER-executable could not be set to a relative
path. This is useful when a checker (compatible substitute or wrapper)
is included with the project.

The setting didn't work because the relative path was passed as-is to
executable-find, which returns nil for inputs containing a directory
component.

Address this by introducing flycheck-default-executable-find, a new
function which checks whether its input has a directory component
before forwarding to executable-find, and update
flycheck-executable-find's default value to point to
flycheck-default-executable-find instead of executable-find.

As its default is being changed, also bump flycheck-executable-find's
package-version to 32