Implement a checker for the golang `staticcheck` tool
This commit implements `go-staticcheck`, a checker wrapping the `staticcheck`
tool. `staticcheck` is the successor to `megacheck`, which is now
deprecated.

This commit includes:

1. A checker definition for `go-staticcheck`.
2. The introduction of a new variable, `flycheck-go-version`, which can be
passed to `staticcheck` to specify the version of the Go compiler to perform
checks compatible with.
3. A test for `staticcheck`.
4. Updates to appropriate documentation.
5. Updates to other Golang syntax checkers to prefer falling back to
`staticcheck` before `megacheck`.

The documentation pass includes:

1. `rst` docs for `go-staticcheck` itself.
2. An update to the documentation for `megacheck`, specifying that it's
deprecated.
3. Documentation for the new `flycheck-go-version` variable.
4. Updating the definition of `flycheck-go-build-tags` to correctly declare the
full list of checkers that can use it.