Distinguish automatically disabled checkers
This commit does the following in order to resolve #1542:

1. Add a variable `-automatically-disabled-checkers` which distinguish
between manually disabled checkers and automatically disabled ones.
This allows us to reset the automatically disabled status without
overwriting user preference from `-disabled-checkers`.

2. Add a function `-reset-enabled-checker` to reset the enabled
or automatically disabled status of a checker.  This allows us to retest
the `:enabled` predicate for a checker.  This function is also useful
for external plugins that may have knowledge that circumstances that led
to a failed `:enabled` test have changed, and a new check must be
done (see #1542).

Closes #1542.