Fix textlint :enabled being too strict; add :verify
The textlint checker :enabled predicate was not using the same logic as
the one used in its :command.  In particular, the :command was picking a
plugin if the current mode was a derived mode of the plugin-alist, but
the :enabled test was doing an `eq` check against the modes in the
list.

org-mode would fail the :enabled test, even though we could use the
default "text" plugin.

Added a corresponding verification line showing the selected plugin.

Close #1564.