Detect erlang-rebar3 projects with dynamic config
Erlang projects can use a `rebar.config.script` file as an alternative to
`rebar.config` in order to have dynamic configuration. The script file will then
be evaluated by rebar3 and the result used as configuration.

This adds `rebar.config.script` as an alternative to `rebar.config` when
detecting an Erlang rebar3 project.