[feat] Support auto-formatting/fixing eslit buffer code
If a user adds `--fix-dry-run` to `flycheck-eslint-args` or otherwise gets `output` to
be returned in the eslint JSON output, then the buffer being checked will be updated
with the fixed output from the JSON.

The one remaining issue I'm aware of is that if there are fixes in the output then
updating the buffer triggers another run of the checker immediately after.  Is there a
way temporarily disable change tracking while calling `replace-buffer-contents`?

It may also be valuable to abstract this so it can be re-used with other
auto-formatters/fixers.

I couldn't run `$ make check specs unit` as it seems to be dependent on Mac OS X.  I did
as much testing as I could.