Allow setting a source suffix for buffers with no file name
Some checkers require that given file names have a certain
suffix (extension) to work, e.g. when the underlying tool makes a
decision of what to do with the file given its file extension.
Specifying the suffix thus becomes necessary when editing code in
transient buffers created by Emacs, which do not have an associated
file name, such as those created by Org-Babel's org-edit-src-code.

For this purpose, extend the source and source-inplace forms
recognized by flycheck-substitute-argument to allow specifying a
suffix to use for temporary files when buffer-file-name is nil.

Note that the aforementioned org-edit-src-code does create buffers
with a non-nil buffer-file-name in the version (9.1.9) currently
distributed with the latest version of Emacs as of today (26.3), so
this change will not be immediately visible unless the user updates
the org-mode package (e.g. from GNU ELPA).