Respect flycheck-highlighting-mode in flycheck-error-pos
With the previous implementation, if the beginning of an overlay didn't match
the beginning of the error, then next-error would behave surprisingly.

For example, if a buffer contained abCd and an error with no end-column started
on C, we'd highlight abCd but jump to C.  This was particularly bad if there was
another error starting on `b', since we'd skip it when jumping right to `C'.

Arguably, jumping to the beginning of the region inferred using
flycheck-highlighting-mode isn't quite right either, since the error might have
purposefully pointed to `C'.  But we already do this for display in any case,
and the problem only exists for checkers that report a position instead of a
region.

One way to work around this would be to constrain the error region to not start
before the position returned by the checker.  This might be a reasonable
extension to explore in the future.

Closes GH-1779.