Set destination to nil in racket checker feature test (#1848).
Sometimes `raco` command will issue a harmless warning to standard
error (eg. when a racket package is installed both in system directory and user
directory). If `destination` is set to `t` in `flycheck-call-checker-process`,
this warning would be improperly redirected to current buffer, causing file
change and syntax error. Then flycheck will call `raco` to check the buffer, and
the warning will be inserted to current buffer again, leading to an infinite
loop.

Setting `destination` to nil avoids problems. Because when
`flycheck-racket-has-expand-p` is called, we don't need `raco`'s output.