isolate the snippet of source code used in formatted error messages
lest bidi formatting characters cause the error message to be rendered
sdrawkcab. In principle we should probably do this to the filename and
even the error message, since those could be suspect as well. As
ordinary RTL text doesn’t cause any problem, it is in practice not so
important. Note also that if flycheck is to be localized, this format
string must be localized as well (or the localization system must
support such substitutions directly).

Fixes #1919.