Enable *-stylelint to work on remote files by using temporary file
The original version used stdin instead of a temporary file (as with
javascript-eslint). However, behavior of eslint and stylelint is
different in terms of stdin processing.

Corresponding CLI commands for both tools are:
eslint:
eslint --format=json --stdin --stdin-filename FILENAME-TO-ASSIGN-TO-STDIN

stylelint:
stylelint --formatter json --stdin-filename FILENAME-TO-ASSIGN-TO-STDIN
or
stylelint --formatter json --stdin --stdin-filename FILENAME-TO-ASSIGN-TO-STDIN

eslint waits for stdin EOF and keeps itself alive, while stylelint
does not wait and may exit before input comes in.

When stylelint does not receive stdin,
it emits its usage message (when --stdin is not specified)
or outputs "empty source" lint message (when --stdin is specified).

For local files, the original version was able to pass input with
`flycheck-process-send-buffer' before stylelint process exited.
For remote files, stylelint process exited before
`flycheck-process-send-buffer' called.

This commit changes *-styelint commands so that they use a temporary
file (instead of stdin) as their input to avoid the above situation on
remote.