Add a file-local-name wrapper function that accepts nil argument
To add support for remote files, some occurrences of filenames had to
be converted to local path on remote (i.e., without
"/method:user@host:" part).
However, these filenames could be nil, while file-local-name signals
an error when its argument is nil.

This led to test failure for scss-lint and others.
(Test command: "make LANGUAGE=scss integ")

This commit adds a wrapper function of file-local-name that accepts
nil as its argument and return it as it is.