Return local home when buffer-file-name returns nil
In `flycheck-locate-config-file-home', file-remote-p is used to
determine if the context is local or remote.
Because file-remote-p does not accept nil,
(file-remote-p (buffer-file-name))
may cause an error.

It led to unit test failure:
   FAILED  flycheck-locate-config-file/existing-file-in-home-directory
   FAILED  flycheck-locate-config-file/existing-file-in-parent-directory
   FAILED  flycheck-locate-config-file/not-existing-file

This commit adds a nil-check for buffer-file-name.
When buffer-file-name returns nil, the context is assumed to be local.